+++ tests/unit/FirewallTest.php
@@ -179,12 +179,18 @@ class FirewallTest extends Common
         $this->mapper->create($this->cfg->getExtendedBanTable(), $data);
         $this->mapper->create($this->cfg->getWhitelistTable(), $data);
         $this->mapper->create($this->cfg->getWhitelistTable(), $data);
+        $this->mapper->create($this->cfg->getFailCountTable(), $data);
+        $this->mapper->create($this->cfg->getFailCountTable(), $data);
+        $this->mapper->create($this->cfg->getBlockCountTable(), $data);
+        $this->mapper->create($this->cfg->getBlockCountTable(), $data);
 
         $elapsedTime = 2;
 
         $this->cfg->setTempBanSeconds($elapsedTime);
         $this->cfg->setExtendedBanSeconds($elapsedTime);
         $this->cfg->setWhitelistSeconds($elapsedTime);
+        $this->cfg->setFailCountSeconds($elapsedTime);
+        $this->cfg->setBlockCountSeconds($elapsedTime);
 
         sleep(2); // value below $elapsedTime should fail this test
 
@@ -200,6 +206,14 @@ class FirewallTest extends Common
 
         // test if IPs are flushed from WhitelistTable
         $rows = $this->mapper->findAll($this->cfg->getWhitelistTable(), $this->where);
+        $this->assertCount(0, $rows);
+
+        // test if IPs are flushed from FailCountTable
+        $rows = $this->mapper->findAll($this->cfg->getFailCountTable(), $this->where);
+        $this->assertCount(0, $rows);
+
+        // test if IPs are flushed from BlockCountTable
+        $rows = $this->mapper->findAll($this->cfg->getBlockCountTable(), $this->where);
         $this->assertCount(0, $rows);
     }
 
@@ -247,6 +261,47 @@ class FirewallTest extends Common
         $this->firewall->flushWhitelisted();
 
         $rows = $this->mapper->findAll($this->cfg->getExtendedBanTable(), $this->where);
+        $this->assertCount(0, $rows);
+    }
+
+    /**
+     * test flushIp
+     */
+    public function testFlushIp()
+    {
+        // seed
+        $this->mapper->create($this->cfg->getBlockCountTable(), $this->data);
+        $this->mapper->create($this->cfg->getExtendedBanTable(), $this->data);
+        $this->mapper->create($this->cfg->getFailCountTable(), $this->data);
+        $this->mapper->create($this->cfg->getTempBanTable(), $this->data);
+        $this->mapper->create($this->cfg->getWhitelistTable(), $this->data);
+
+        $this->firewall->flushIp($this->testIp);
+
+        // test if IPs are flushed from BlockCountTable
+        $rows = $this->mapper->findAll($this->cfg->getBlockCountTable(), $this->where);
+        $this->assertCount(0, $rows);
+
+        // test if IPs are flushed from ExtendedBanTable
+        $rows = $this->mapper->findAll($this->cfg->getExtendedBanTable(), $this->where);
+        $this->assertCount(0, $rows);
+
+        // test if IPs are flushed from FailCountTable
+        $rows = $this->mapper->findAll($this->cfg->getFailCountTable(), $this->where);
+        $this->assertCount(0, $rows);
+
+        // test if IPs are flushed from TempBanTable
+        $rows = $this->mapper->findAll($this->cfg->getTempBanTable(), $this->where);
+        $this->assertCount(0, $rows);
+
+        // test if IPs is NOT flushed from WhitelistTable
+        $rows = $this->mapper->findAll($this->cfg->getWhitelistTable(), $this->where);
+        $this->assertCount(1, $rows);
+
+        $this->firewall->flushIp($this->testIp, true);
+
+        // test if IPs are flushed from WhitelistTable
+        $rows = $this->mapper->findAll($this->cfg->getWhitelistTable(), $this->where);
         $this->assertCount(0, $rows);
     }
 }