+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 1.1.0
+
+- Add flushIp() service method.
+- Auto-expire entries in failCount table.
+- Add setFailCountSeconds() config method.
+- Auto-expire entries in blockCount table.
+- Add setBlockCountSeconds() config method.
+
 ## 1.0.0
 
 - Release first version.